Needs augmenting paths, and metod returns needs to be evaluated.
O2 needs to be changed. The algorithm runs in O(n) time which might be reduced some way.
(Maybe) A seach function that iterates through the spaces furthest away. This would be a radius search which could be either really bad, or really effective. WHY? Because if the array only have few values but a large 2d array, then the search will travese more than the initial array which is n. The run time would be O(n^2)

If i implement this, my algorithm teacher might kill me.